requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
flake8
gunicorn
django-heroku
pytest
pytest-django
model_bakery
